requirements.txt
# Core Agent Framework
autogen-agentchat>=0.2.0
autogen-core>=0.2.0
autogen-ext>=0.2.0

# Google Gmail API
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1
google-api-python-client>=2.100.0

# Azure OpenAI (for agents)
openai>=1.3.0
azure-identity>=1.14.0

# Data Processing
python-dotenv>=1.0.0
pydantic>=2.4.0
pydantic-settings>=2.0.0

# Utilities
python-dateutil>=2.8.2
beautifulsoup4>=4.12.0
html2text>=2020.1.16

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
pytest-mock>=3.11.0

# Development
jupyter>=1.0.0
notebook>=7.0.0
ipykernel>=6.25.0
black>=23.9.0
flake8>=6.1.0
mypy>=1.5.0

# Type stubs
types-python-dateutil>=2.8.0
